backend/requirements.txt
alembic==1.9.2
Flask
gunicorn
Flask_Migrate==4.0.3
flask_sqlalchemy==3.0.3
pendulum==2.1.2
python-dotenv==0.21.1
requests==2.25.1
SQLAlchemy==2.0.0
